tidy up and example
