+++ tests/create_contract/src/lib.rs
@@ -1,5 +1,5 @@
 #![no_std]
-use soroban_sdk::{contractimpl, Bytes, BytesN, Env};
+use soroban_sdk::{contractimpl, Bytes, BytesN, CurrentNamespace, Env};
 
 pub struct Contract;
 
@@ -8,6 +8,7 @@ impl Contract {
     // Note that anyone can create a contract here with any salt, so a users call to
     // this could be frontrun and the same salt taken.
     pub fn create(e: Env, c: Bytes, s: BytesN<32>) {
-        e.create_contract_from_contract(c, s);
+        let _contract_id = e.contract_id(CurrentNamespace, &s);
+        let _contract_id = e.deployer(CurrentNamespace).deploy_wasm(&s, c);
     }
 }
